fix leak at preview: implement cancelation flow
(STR: very QUICKLY open and close settings modal on video tab.)

this time, the issue was for camera preview,
which is usually more slower to load compared to microphone.

too many start/stop during such slow action provoked
multiple starts to be active while the last cleanup was
getting forgotten by the react cycle.

eventually, as I could observe, the stream would get terminated
anyway by the garbage collector, but I didnt want to rely on that.

the current lock solution is good enough to narrow the issue
to a single missing `stop` (cause the `start` itself was well locked),
but there wasnt anything else to handle this last `close` in
case it was missed by the last `useEffect` cleanup.

an easy alternative would be a setTimeout when the context unmounts,
but I didnt to rely on such sloppy approach, cause the camera might
take more than the whole timeout to get ready or even never get ready,
thus forcing me to add recursive attempts to close or just give up.