wip: start refactoring, user status part of call state
This data actually doesnt matter much for the listing, its actually
a local info of what UI to show.

in the previous way it became a blocker, because I had to check in two different
places (users list and call status) to decide what to show to the user.

listing was managed by joining flow while call status was managed by
creation flow, thus breaking the concept of single source of truth that
redux wants.